- Fixed a small datatype error in release build
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1833 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/Common/AESEncryption.cpp
@@ -127,7 +127,7 @@ void AESEncryption::shiftRow(uint8_t *state, uint8_t nbr) {
 }
 
 void AESEncryption::shiftRows(uint8_t *state) {
-	int32_t i;
+	int8_t i;
 	for (i = 0; i < 4; i++) {
 		shiftRow(state + i * 4, i);
 	}
@@ -135,7 +135,7 @@ void AESEncryption::shiftRows(uint8_t *state) {
 
 
 void AESEncryption::addRoundKey(uint8_t *state, uint8_t *roundKey) {
-	int32_t i;
+	int8_t i;
 	for (i = 0; i < 16; i++) {
 		state[i] = state[i] ^ roundKey[i];
 	}
